# docs/openapi.yml
openapi: 3.1.0
info:
  title: Tunecamp REST API
  version: 2.0.0
  description: |
    The Tunecamp API provides programmatic access to the platform's music library, artists, albums, tracks,
    releases, playlists, and administrative functions. It supports the decentralized music ecosystem via
    Zen and ActivityPub integrations.

    ## Authentication
    Public endpoints (catalog browsing, streaming) require no authentication.
    User endpoints require a Bearer token from `/api/auth/login`.
    Admin endpoints require a Bearer token with an admin role.

    ## Subsonic API
    For Subsonic/OpenSubsonic client compatibility, use the `/rest` endpoint.
    See [SUBSONIC.md](./SUBSONIC.md) for the full endpoint reference.
servers:
  - url: /api
    description: Tunecamp API base path
paths:
...
  /users/register:
    post:
      summary: User Registration
      tags: [Users]
      description: Registers a new user. Prioritizes Zen identity linking.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                pubKey:
                  type: string
                  description: The user's Zen SEA public key.
                signature:
                  type: string
                  description: Proof-of-possession signature.
      responses:
        '200':
          description: Registration successful.
        '409':
          description: Username already taken.

  # ── Catalog (Public) ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...